GCM_Sales_Chatbot/requirements.txt
flask==2.0.1
python-docx==0.8.11
langchain==0.0.267
python-dotenv==0.19.0
groq==0.4.2
faiss-cpu==1.7.4
tiktoken==0.4.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-multipart==0.0.6
pydantic==1.10.12
typing==3.7.4.3
colorama==0.4.6
tqdm==4.65.0
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
torch==2.0.1 